Setup initial node environment and Strava module
Can now make calls as an authenticated user (me) to Strava API and retrieve data. Currently pulls an array of activities when going to endpoint '/activities' and returns an object with some basic activity info for a user. Currently just a proof-of-concept for data retrieval, not any particular design philosophy, will likely change.